Introduce boundary module, make capture/replay generic in it
